fix(S07): update sort logic
